JAMES-3012 Moving BasicFeature to a generic maven module
