+++ server/protocols/jmap-draft-integration-testing/cassandra-jmap-draft-integration-testing/pom.xml
@@ -225,7 +225,7 @@
                     <reuseForks>true</reuseForks>
                     <!-- Fail tests longer than 2 hours, prevent form random locking tests -->
                     <forkedProcessTimeoutInSeconds>7200</forkedProcessTimeoutInSeconds>
-                    <groups combine.self="override">org.apache.james.jmap.categories.CassandraAndElasticSearchCategory,org.apache.james.jmap.categories.BasicFeature,org.apache.james.jmap.categories.EnableCucumber</groups>
+                    <groups combine.self="override">org.apache.james.jmap.categories.CassandraAndElasticSearchCategory,org.apache.james.junit.categories.BasicFeature,org.apache.james.jmap.categories.EnableCucumber</groups>
                 </configuration>
             </plugin>
         </plugins>
